sched: cpufreq: sysfs knob for capacity_margin
This commit makes the capacity_margin knob configurable from user-space.
By tinkering with this knob, cpufreq scaling of frequency can be made
more aggressive. Allowed values for this knob range from 0 to 1024. By
default, it has the value of 256 which sets the frequency scaling factor
to be 1.25

Bug 2106745

Change-Id: Iaed5914eaa9c091aa6635f1595ca331df726dd17
Signed-off-by: Waqar Ali <waqara@nvidia.com>
Signed-off-by: Somdutta Roy <somduttar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1810752
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>