drivers: tegra: Reset all clk modules in SOR PD
This change includes the following clk module to be reset as a part of
the powergating/unpowergating sequence for SOR:
- TEGRA210_CLK_DSIA
- TEGRA210_CLK_DPAUX
- TEGRA210_CLK_DPAUX1

The TEGRA210_CLK_DISP1 clk is removed from the reset pool for the SOR
power domain (PD). This module is a part of the DISA PD and will be
reset separately as a part of the DISA powergate/unpowergate sequence.

Based on the SOR powering sequence, it is required to reset all the
clk modules that belong to a particular PD. For SOR these include:
- SOR0
- SOR1
- DPAUX
- DPAUX1
- DSIA
- DSIB
- MIPICAL

The earlier code ignored DPAUX and led to timeouts for I2C transactions
(which use the DPAUX pad) with certain TVs and devices.

Bug 2356564

Change-Id: I9bb1cb355ed03ff68c6f241b2e5a0cdcca5f7d2b
Signed-off-by: Abhiroop Kaginalkar <akaginalkar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1923895
Reviewed-by: Santosh Galma <galmar@nvidia.com>
Reviewed-by: Prafull Suryawanshi <prafulls@nvidia.com>
Tested-by: Prafull Suryawanshi <prafulls@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>