Add NvIDS in brcmfmac driver
NV IDS is an automatic reporting system based on the predefined set of
problems that can happen on user device. This system is currently
targeting communication stack to detect an issue and process match
the issue with the predefined set of strings and then pattern matching
with these strings. Once a certain problem is identified, it will issue
a bugreport along with more logs to a local directory
or Nvidia Ranger server.

Link:
https://wiki.nvidia.com/wmpwiki/index.php/Communication_software/IDS

NvTag: NvConnectivityDebug
NvFeature: NvIDS

Bug 200465854

Change-Id: I868fb736d4d65b1900ac57469a7b85ceda9a9aa4
Reviewed-on: https://git-master.nvidia.com/r/1990949
Reviewed-by: Mohan Thadikamalla <mohant@nvidia.com>
Tested-by: Mohan Thadikamalla <mohant@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Krishna Thota <kthota@nvidia.com>