memory: tegra: make emc driver to suspend late
PM framework adds devices to its suspend list based on the order
in which the devices are created during boot. Earlier, this used
to happen as part of init calls of various modules, which had a
specific heirarchy based on the type of init call. But, with the
introduciton of device tree, devices are created as per the order
of device nodes listed in DT. This order doesn't necessarily
specify all dependencies between devices. And hence, the order in
which devices are added to suspend list is not guaranteed to take
care of dependencies.

Given this, EMC driver might get suspended before any of its client
drivers because devices that request memory bandwidth are not always
specified in after emc node in DT. And when EMC driver suspends, it
sets emc clock to 204Mhz irrespective of any client requests causing
bandwidth issues to clients.
To avoid such a scenario, updated EMC driver pm_ops to late-suspend
and early-resume.

bug 200488631

Change-Id: I66a936b7436fc112a1d5cb697670900533656bfc
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2013302
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Shu Zhong <shuz@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>