pinctrl: tegra: Save and restore pinctrl state for GPIO requests
For Tegra186 and Tegra194, when a GPIO is requested, the pinmux for the
associated pin is configured as a GPIO after the direction of the GPIO
is configured. When the GPIO is freed the pinmux for the pin is
reconfigured as an SFIO regardless of whether the pin was configured as
a GPIO or SFIO before it was requested. It should be noted that Tegra210
does not behave this way and restores the state of the pin on freeing
the GPIO.

Update the Tegra pinctrl driver so that the state of the pinmux is
restored on freeing a GPIO. This is achieved by calling the function
pinctrl_gpio_save_config() when a GPIO is requested to save the pinmux
state and calling the function pinctrl_gpio_restore_config() on freeing
the GPIO to restore the pinmux state.

Bug 2414131

Change-Id: Iec8e02007cefc8c5eff57be731007595080ed8cb
Signed-off-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2155336
Reviewed-by: Stephen Warren <swarren@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>