soc/tegra: pmc: add reset reason for PMIC watchdog
After enabling PMIC watchdog in T210 devices from Bug-200577867 when reset
occur due to PMIC watchdog, It should be print as a PMIC_WATCHDOG_POR in tegra
reset reason.

In Mb1 bootloader, NVERC register from PMIC is stored into the SCRATCH203
register. Adding support to read it and decide that whether Power on Reset
is actual POR or because of the PMIC watchdog.

If POR is because of the PMIC watchdog then print tegra reset reason as
a PMIC_WATCHDOG_POR

Bug 200624822

Signed-off-by: Ankit Patel <anpatel@nvidia.com>
Change-Id: I3a4031f13227c628e02bc3050f84e831fa437ac9
Reviewed-on: https://git-master.nvidia.com/r/c/linux-4.9/+/2360134
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>