gpio: tegra: Fix input setup for GPIO IRQs
If a GPIO is used as an interrupt that is configured in device-tree by
using the 'interrupts' property, then if the input-enable is not set for
the pin, no interrupts will be received. The GPIO is correctly
configured as an input in the GPIO controller, but not at the pin level.
Fix this by calling tegra_gpio_direction_input() from within the
tegra_gpio_irq_set_type() function to ensure that the function
pinctrl_gpio_direction_input() is called to enable the input at the pin
level.

Bug 200637207

Change-Id: If89cb3cfcfad0971f2b5557803f06a48f1f0fa33
Signed-off-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-4.9/+/2395368
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>