bm92txx: Add support for plug-in wake up
In case the device is in suspend, this allows initialization of a new connection on wake-up.

It checks if there's a new contract and continues with the process.

In case the charger is PD, we executed too fast and there's no PDO alert, it waits for it.
This ensures that the pdo check can succeed.

(cherry picked from commit bff04a3c9858f2d4fb08ca708658f5f4ad18e0cf)