thermal: max77620: change MAX77620_NORMAL_OPERATING_TEMP to 50C
MAX77620_NORMAL_OPERATING_TEMP is a dummy temperature and was chosen
as 100C because thermal monitor of max77620 can only tell temperature
is higher or lower than 120C (and 140C). However, the current value
often confuses users because 100C is too high for most thermal design
cases. So lowering that value to 50C so as not to make users wonder if
something goes wrong with thermal solutions.

Bug 200679123

Signed-off-by: Leon Yu <leoyu@nvidia.com>
Change-Id: If2960f06f1f53c7911562d9524b662fa5404d111
Reviewed-on: https://git-master.nvidia.com/r/c/linux-4.9/+/2503014
Reviewed-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>