arm64: tegra: Enable actmon
There are cases where CPU usage is low but EMC pressure is high, either because of GPU or because of dma engines.

While CPU based EMC scaling guaranties minimum EMC frequencies, it's not enough because it can completely starve these functions.

So enable actmon as we heavily rely on it setting proper EMC frequencies based on EMC bandwidth consumed.