bq2419x-charger: Add charge limits based on soc
This adds support for battery life protection by regulating charging based on battery levels.

Enabling this with a proper value, sets limits inside -+5%.
For example a level of 55 will enable charging if battery is < 50% and disable it if > 60%.
Inside that region, charging state doesn't change forcibly.

The min/max limits are 45 and 85 which are translated to 40% and 90%. If wrong value is set then the charging is enabled.