regulator: max8973: change MAX77621_NORMAL_OPERATING_TEMP to 55C
MAX77621_NORMAL_OPERATING_TEMP is a dummy temperature and was chosen
as 100C because thermal monitor of max77621 can only tell temperature
is higher or lower than 120C (and 140C). However, the current value
often confuses users because 100C is too high for most thermal design
cases. So lowering that value to 55C so as not to make users wonder if
something goes wrong with thermal solutions.