xhci: tegra: fix a detection race at boot
Before the driver would register the extcon detections on probing.

In platforms that rely on firmware to be loaded that would cause anything
that was plugged between that time and the forced extcon check to miss any
plug-in event.

That could be fixed by correcting the forced checks and re-issue an event for
already plugged ports or by moving the extcon register on probe finish.

The latter was chosen for simplicity and because acting on events is instant.