power: supply: bq2419x: add CE pin support
Some boards utilize CE pin to fine control charging.

Utilize it for any charging disable function.

This also allows for restoring charging in case it's disabled and board hangs,
as all GPIOs are reset on boot.

An example is charging being disabled in sleep and a shut off or hang could
happen and charging would remain disabled. And if low battery monitor is active,
the device can't be woken until bq chip is reset.