input: joycon: improve sio calibration
The expected calibration on `sio-stick-cal-l` and `sio-stick-cal-r`
is now the following:
X- range, X center, X+ range, Y- range, Y center, Y+ range.
So the expected raw calibration is expected.
Additionally, -+ denotes the stick orientation and not the actual value.
That is handled internally.

Finally, `sio-stick-cnt-off-l` and `sio-stick-cnt-off-r` were added to move the
center away from original calibration if needed.
Sio seems to have a ~0x100 offset on left Y and right X.

And increase deadzone to 5%.