hid: nintendo: consider baudrate change non-fatal
Every official controller that uses nxpad protocol is usb2uart so that command
is used to change the internal uart rate from 1 mbps to 3 mbps.

Some bad clones that are usb only, incorrectly, do not emulate that command and
timeout. So do not fail if it's not supported and only do handshake again if
succeeded.